API

API

Constants

Utilities

Constants

SuiteSparseMatrixCollection.ssmc — Constant

Main database.

SuiteSparseMatrixCollection.ssmc_formats — Constant

Formats in which matrices are available.

Utilities

SuiteSparseMatrixCollection.ssmc_matrices — Function

ssmc_matrices(group, name)

Return a DataFrame of matrices whose group contains the string group and whose name contains the string name.

ssmc_matrices(name)
ssmc_matrices("", name)

Return a DataFrame of matrices whose name contains the string name.

ssmc_matrices(group, "")

Return a DataFrame of matrices whose group contains the string group.

Example: ssmc_matrices("HB", "bcsstk").

SuiteSparseMatrixCollection.fetch_ssmc — Function

 fetch_ssmc(group::AbstractString, name::AbstractString; format="MM")

Download the matrix with name name in group group. Return the path where the matrix is stored.

 fetch_ssmc(matrices; format="MM")

Download matrices from the SuiteSparseMatrixCollection. The argument matrices should be a DataFrame or DataFrameRow. An array of strings is returned with the paths where the matrices are stored.