Reference

​

Contents

​

Reference

Contents

Index

​

Index

​

SuiteSparseMatrixCollection.ssmc

SuiteSparseMatrixCollection.ssmc_formats

SuiteSparseMatrixCollection.fetch_ssmc

SuiteSparseMatrixCollection.fetch_ssmc

SuiteSparseMatrixCollection.ssmc_matrices

​

SuiteSparseMatrixCollection.ssmc — Constant

Main database.

SuiteSparseMatrixCollection.ssmc_formats — Constant

Formats in which matrices are available.

SuiteSparseMatrixCollection.fetch_ssmc — Method

 fetch_ssmc(group::AbstractString, name::AbstractString; format="MM")

Download the matrix with name name in group group. Return the path where the matrix is stored.

SuiteSparseMatrixCollection.fetch_ssmc — Method

 fetch_ssmc(matrices; format="MM")

Download matrices from the SuiteSparseMatrixCollection. The argument matrices should be a DataFrame or DataFrameRow. An array of strings is returned with the paths where the matrices are stored.

SuiteSparseMatrixCollection.ssmc_matrices — Method

ssmc_matrices(group, name)

Return a DataFrame of matrices whose group contains the string group and whose name contains the string name.

ssmc_matrices(name)
ssmc_matrices("", name)

Return a DataFrame of matrices whose name contains the string name.

ssmc_matrices(group, "")

Return a DataFrame of matrices whose group contains the string group.

Example: ssmc_matrices("HB", "bcsstk").