Reference

Contents

Reference

Contents

Index

Index

SuiteSparseMatrixCollection.ssmc_formats

SuiteSparseMatrixCollection.delete_all_ssmc

SuiteSparseMatrixCollection.delete_ssmc

SuiteSparseMatrixCollection.fetch_ssmc

SuiteSparseMatrixCollection.fetch_ssmc

SuiteSparseMatrixCollection.installed_ssmc

SuiteSparseMatrixCollection.ssmc_db

SuiteSparseMatrixCollection.ssmc_matrices

SuiteSparseMatrixCollection.ssmc_formats — Constant

Formats in which matrices are available.

SuiteSparseMatrixCollection.delete_all_ssmc — Method

delete_all_ssmc()

Remove all matrices from the SuiteSparseMatrixCollection.

SuiteSparseMatrixCollection.delete_ssmc — Function

delete_ssmc(name::AbstractString, group::AbstractString, format = "MM")

Remove the matrix with name name in group group and format format.

SuiteSparseMatrixCollection.fetch_ssmc — Method

 fetch_ssmc(group::AbstractString, name::AbstractString; format="MM")

Download the matrix with name name in group group. Return the path where the matrix is stored.

SuiteSparseMatrixCollection.fetch_ssmc — Method

 fetch_ssmc(matrices; format="MM")

Download matrices from the SuiteSparseMatrixCollection. The argument matrices should be a DataFrame or DataFrameRow. An array of strings is returned with the paths where the matrices are stored.

SuiteSparseMatrixCollection.installed_ssmc — Method

installed_ssmc()

Return a vector of tuples (group, name, format) of all installed matrices from the SuiteSparseMatrixCollection.

SuiteSparseMatrixCollection.ssmc_db — Method

ssmc_db(; verbose::Bool=true)

Load the database of the SuiteSparseMatrixCollection. A summary of the statistics available for each matrix can be found at https://www.cise.ufl.edu/research/sparse/matrices/stats.html.

SuiteSparseMatrixCollection.ssmc_matrices — Method

ssmc_matrices(ssmc, group, name)

Return a DataFrame of matrices whose group contains the string group and whose name contains the string name.

ssmc_matrices(ssmc, name)
ssmc_matrices(ssmc, "", name)

Return a DataFrame of matrices whose name contains the string name.

ssmc_matrices(ssmc, group, "")

Return a DataFrame of matrices whose group contains the string group.

Example: ssmc_matrices(ssmc, "HB", "bcsstk").